fix Font point size fields in Properties panel only show/accept wholFont point size fields in Properties panel only show/accept whole numbers #16764

+++ src/inspector/models/text/textsettingsmodel.cpp
@@ -111,8 +111,8 @@ void TextSettingsModel::loadProperties()
     m_fontStyle->setIsEnabled(true);
 
     loadPropertyItem(m_fontSize, [](const QVariant& elementPropertyValue) -> QVariant {
-        return elementPropertyValue.toInt() == mu::engraving::TextBase::UNDEFINED_FONT_SIZE
-               ? QVariant() : elementPropertyValue.toInt();
+        return elementPropertyValue.toDouble() == mu::engraving::TextBase::UNDEFINED_FONT_SIZE
+               ? QVariant() : elementPropertyValue.toDouble();
     });
 
     m_fontSize->setIsEnabled(true);
